# [Optimization] Moving get snapshot requests to listener based async calls 