--- a/plugins/repository-azure/licenses/azure-json-1.0.1.jar.sha1
+++ b/plugins/repository-azure/licenses/azure-json-1.0.1.jar.sha1
@@ -1,0 +1,1 @@
+abdfdb0c49eebe75ed8532d047dea0c9f13c30ac