… CHANGELOG.md
… plugins/repository-azure/licenses/azure-core-1.34.0.jar.sha1
… plugins/repository-azure/licenses/azure-core-1.39.0.jar.sha1
… server/src/main/java/org/opensearch/index/shard/RemoteStoreRefreshListener.java
… server/src/test/java/org/opensearch/threadpool/ScalingThreadPoolTests.java
… test/fixtures/hdfs-fixture/build.gradle
… test/framework/src/main/java/org/opensearch/snapshots/mockstore/MockRepository.java